Added a minor fix to the angle finder function.

+++ R2D2/Angle_Finder.ino
@@ -1,3 +1,3 @@
 float angle_finder(int x_value) {
-  return atan((x_value - 160) / FOCAL_LENGTH);
+  return atan(static_cast<float>(x_value - 160) / FOCAL_LENGTH);
 }
